requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
psycopg==3.1.9
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
